# kanabi/static/swagger.yml
openapi: 3.0.1
info:
  title: Kanabi
  description: Spreadsheet ingestor and data management tool for the Department of
    Community and Civic Life's cannabis licensing program.
  contact:
    name: Project Lead
    email: andrew13haley@gmail.com
  version: 1.0.0
servers:
- url: https://localhost:443/
paths:
...
  /admin/{mode}:
    post:
      summary: Update user profile information
      description: Allows an admin-level user to update a given user's profile.
      parameters:
        - name: mode
          in: path
          description: |
            Operation to perform. Currently supported operations are: \
            `changepassword`: change the user's password. Requires `password` argument in the form. \
            `changename`: change the user's `name` value. Required argument: `name` \
            `changeemail`: change the user's email. Required argument: `new_email` \
            `removeuser`: de-register the listed user. The user is removed from the database. \
            `makeadmin`: register the listed user as an admin. \
            `removeadmin`: remove the listed user's admin permissions. \
            `makeeditor`: sets the user to editor status (default) \
            `removeeditor`: sets the user to read-only status
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                new_email:
                  type: string
              required:
                - email
                - password
                - name
      responses:
        200:
          description: Operation successful.
          content:
            application/json: {}
        400:
          description: Required parameter not provided, or request invalid.
          content:
            application/json: {}
        403:
          description: User is not logged in as an admin.
          content:
            application/json: {}
    get:
      summary: Retrieve data needed for user administration.
      description: Allows an admin-level user to retrieve privileged user data, such as a listing of all users.
      parameters:
        - name: mode
          in: path
          description: |
            Operation to perform. Currently supported operations are: \
            `changepassword`: change the user's password. Requires `password` argument in the form. \
            `changename`: change the user's `name` value. Required argument: `name` \
            `changeemail`: change the user's email. Required argument: `email` \
            `removeuser`: de-register the active user. The user is removed from the database.
          required: true
          schema:
            type: string
      responses:
        200:
          description: Request successful
          content:
            application/json: {}
        400:
          description: Required parameter not provided, or request invalid
          content:
            application/json: {}
        403:
          description: User is not logged in as an admin.
          content:
            application/json: {}
  /edituser/{mode}:
    post:
      summary: Change user profile
      description: Allows a user to edit their own profile information. An admin may also edit another user's profile.
      parameters:
        - name: mode
          in: path
          description: Operation to perform. Currently supported operations are 'changepassword', 'changename', 'changeemail', and 'removeuser'. See '#/paths/admin/' for operation descriptions.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        200:
          description: Request successful.
          content:
            application/json: {}
        400:
          description: Request failed.
          content:
            application/json: {}
  /makeadmin:
    post:
      summary: Create initial admin account
      description: Creates the first admin account. Fails if another user exists with admin permissions. Used in
        setting up the database, so the initial user can manage subsequent users.
      requestBody:
        description: Email and password for this user. Name may also be supplied.
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
              required:
                - email
                - password
      responses:
        200:
          description: Account successfully created
          content:
            application/json: {}
        400:
          description: An admin account already exists.
          content:
            application/json: {}

# Data Management
  /list:
    get:
      summary: Displays the contents of a table
      description: Returns a list of the table's records
      parameters:
      - name: table
        in: query
        description: Name of table to display
        required: true
        schema:
          type: string
      responses:
        200:
          description: Contents of table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/table'
        400:
          description: Table name not supplied
          content:
            application/json: {}
        404:
          description: Named table does not exist
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json: {}
    post:
      summary: Filter table data
      description: Returns a filtered subset of data based on a POSTed query in JSON
        format.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/query_request'
      responses:
        200:
          description: Requested data
          content:
            application/json: {}
        400:
          description: Query could not be processed
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
  /load:
    put:
      summary: Inserts a row
      description: Upload a single row and insert it into the named table.
      parameters:
      - name: table
        in: query
        description: Name of table to insert into
        required: true
        schema:
          type: string
      requestBody:
        required: true
        description: JSON object of row to insert
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/table_row'
      responses:
        200:
          description: Row insertion successful
          content:
            application/json: {}
        400:
          description: Table name not supplied
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
    post:
      summary: Uploads a file
      description: Add the contents of a file to the database. Must be a .xls or .xlsx file.
      parameters:
      - name: table
        in: query
        description: Name of table to insert data into
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
            schema:
              type: string
              format: binary
          application/vnd.ms-excel:
            schema:
              type: string
              format: binary
      responses:
        200:
          description: File successfully consumed
          content:
            application/json: {}
        400:
          description: File name not supplied
          content:
            application/json: {}
        404:
          description: File could not be located
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
  /delete:
    get:
      summary: Delete a row
      description: Delete specified row(s) from specified table
      parameters:
        - name: table
          in: query
          required: true
          description: Name of the table to delete from
          schema:
            type: string
        - name: row
          in: query
          required: true
          description: Integer of row ID to be deleted. Multipe IDs are separated by '+', e.g. "row=1+2"
          schema:
            type: string
      responses:
        200:
          description: Row(s) successfully deleted
          content:
            application/json: {}
        400:
          description: No table/row specified in query
          content:
            application/json: {}
        404:
          description: Table doesn't exist or row input invalid
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
  /export:
    get:
      summary: Export data as a CSV file
      description: Save a table's data into a CSV file
      parameters:
        - name: table
          in: query
          required: true
          description: The name of the db table to be exported
          schema:
            type: string
      responses:
        200:
          description: File containing table contents in csv format
          content:
            text/plain; charset=utf-8: {}
        400:
          description: Table name not specified
          content:
            application/json: {}
        404:
          description: Table does not exist
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
  /update:
    post:
      summary: Update record in table
      description: Update the contents of the table. Currently only supports the intake table.
      parameters:
        - name: content_type
          in: header
          required: true
          description: String containing encoding structure of input
          schema:
            type: string
          style: simple
        - name: row
          in: query
          required: true
          description: Index of row to update
          schema:
            type: integer
      responses:
        200:
          description: Successful update of record
          content:
            application/json: {}
        400:
          description: No column/row specified, or invalid data supplied
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
  /metadata:
    get:
      summary: View file metadata
      description: Displays the contents of the metadata table.
      responses:
        200:
          description: Contents of metadata table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/metadata'
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
  /restore:
    put:
      summary: restore row from archive
      description: restore a row from archive table back to original table
      parameters:
        - name: row
          in: query
          required: true
          description: The row id in the archive table to restore
          schema:
            type: integer
      responses:
        200:
          description: successful update of record
          content:
            application/json: {}
        400:
          description: No column/row specified, invalid data specified
          content:
            application/json: {}
        500:
          description: An unexpected error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Standard500ErrorResponse'
components:
  schemas:
...
    query_request:
      type: object
      properties:
        table:
          type: string
        columns:
          type: array
          items:
            type: string
        where:
          $ref: '#/components/schemas/op_block'
      required:
        - table
    and_block:
      type: array
      items:
        $ref: '#/components/schemas/op_block'
      minItems: 2
      maxItems: 2
    or_block:
      type: array
      items:
        $ref: '#/components/schemas/op_block'
      minItems: 2
      maxItems: 2
...
    op_block:
      type: object
      oneOf:
        - $ref: '#/components/schemas/and_block'
        - $ref: '#/components/schemas/or_block'
        - $ref: '#/components/schemas/op'
    table_row:
      oneOf:
        - $ref: '#/components/schemas/intake'
        - $ref: '#/components/schemas/metadata'
        - $ref: '#/components/schemas/reports'
        - $ref: '#/components/schemas/violations'
        # archive
        # txn_history
        # add more here once more tables are defined
    table:
      type: array
      items:
        $ref: '#/components/schemas/table_row'
...
    metadata:
      type: object
      description: Record of metadata captured when a file is uploaded to the database
      properties:
        filename:
          type: string
        creator:
          type: string
        size:
          minimum: 0
          type: integer
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        lastModifiedBy:
          type: string
        title:
          type: string
      required:
        - filename
        - creator
        - size
        - created
        - modified
        - lastModifiedBy
        - title
...
  responses:
    Standard500ErrorResponse:
      description: An unexpected error occured.
      content:
        application/json: {}